# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "200":
          description: Registered and signed in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Me"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Sign in with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Signed in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Me"
...
  /auth/forgot-password:
    post:
      operationId: forgotPassword
      tags: [auth]
      summary: Start a password reset for the given email, if an account exists
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ForgotPasswordInput"
      responses:
        "200":
          description: Reset started (response is intentionally generic to avoid leaking account existence)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ForgotPasswordResult"
  /auth/reset-password:
    post:
      operationId: resetPassword
      tags: [auth]
      summary: Complete a password reset using a reset token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPasswordInput"
      responses:
        "200":
          description: Password updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResetPasswordResult"
...
  /admin/users/{userId}/password-reset:
    post:
      operationId: adminResetUserPassword
      tags: [admin]
      summary: Generate a one-time password reset link for a user (admin-assisted reset, no email sending configured)
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Reset link generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminPasswordResetResult"
...
components:
  schemas:
...
    UserRole:
      type: string
      enum: [user, admin]
    RegisterInput:
      type: object
      required: [email, password, ref]
      properties:
        email: { type: string, format: email }
        password: { type: string, minLength: 8, maxLength: 200 }
        name: { type: string }
        # Invite-only signup: must match an existing user's referralCode.
        # There is no bootstrap exception — the seeded admin account already
        # has one, so it's the root of every invite chain.
        ref: { type: string, minLength: 1 }
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string, minLength: 1, maxLength: 200 }
    ForgotPasswordInput:
      type: object
      required: [email]
      properties:
        email: { type: string, format: email }
    ForgotPasswordResult:
      type: object
      required: [message]
      properties:
        message: { type: string }
    ResetPasswordInput:
      type: object
      required: [token, password]
      properties:
        token: { type: string, minLength: 1 }
        password: { type: string, minLength: 8, maxLength: 200 }
    ResetPasswordResult:
      type: object
      required: [message]
      properties:
        message: { type: string }
...
    Me:
      type: object
      required:
        [
          id,
          email,
          role,
          hasActiveSubscription,
          deviceSlots,
          activeKeyCount,
          balanceKopecks,
          referralCode,
          referralCommissionPercent,
          referralEarningsKopecks,
          referredUserCount,
          referredPayingUserCount,
          referralLinkHost,
          extraTrafficGb,
          trafficLimitExceeded,
          isBanned,
          isTrialSubscription,
          autoRenewFromBalance,
          subscriptionState,
          expiredSubscription,
        ]
      properties:
        id: { type: integer }
        email: { type: string }
        name: { type: ["string", "null"] }
        role:
          $ref: "#/components/schemas/UserRole"
        hasActiveSubscription: { type: boolean }
        # The current access state. "expired" is distinct from "none" so the
        # cabinet can explain why access stopped rather than treating a former
        # subscriber like a user who never selected a plan.
        subscriptionState:
          type: string
          enum: [active, expired, none]
        # Most recent ended subscription when subscriptionState is "expired".
        # Kept nullable for active and never-subscribed users. Hourly
        # subscriptions may have endsAt = null because their billing is
        # balance-driven; their existing balance_exhausted UI remains the
        # user-facing owner of that state.
        expiredSubscription:
          oneOf:
            - $ref: "#/components/schemas/ExpiredSubscription"
            - type: "null"
        currentPlanName: { type: ["string", "null"] }
        subscriptionEndsAt: { type: ["string", "null"], format: date-time }
        currentPlanBillingType:
          oneOf:
            - $ref: "#/components/schemas/PlanBillingType"
            - type: "null"
        hourlyRateKopecks: { type: ["integer", "null"] }
        lastBilledAt: { type: ["string", "null"], format: date-time }
        deviceSlots: { type: integer }
        activeKeyCount: { type: integer }
        balanceKopecks: { type: integer }
        # Effective traffic cap for the current plan's period, in gigabytes —
        # already includes any self-service top-up (extraTrafficGb). Null
        # when there is no active subscription or the plan has no cap.
        trafficLimitGb: { type: ["integer", "null"] }
        # Extra traffic (GB) purchased on top of the plan's base allowance
        # for the current subscription period. 0 if none bought.
        extraTrafficGb: { type: integer }
        # true once the current subscription's period usage has exceeded its
        # effective traffic cap — new VPN key issuance is blocked while this
        # is true (see keyIssuance.ts isTrafficLimitBlocked). Cleared by
        # buying extra traffic or by the subscription renewing.
        trafficLimitExceeded: { type: boolean }
        # Sum of periodUpBytes + periodDownBytes across all of the user's
        # non-revoked VPN keys, for showing usage vs trafficLimitGb.
        periodUsageBytes: { type: integer }
        # Personal invite code, shown as a shareable link in the profile.
        referralCode: { type: string }
        # % of a referral's confirmed subscription payment credited to this
        # user's wallet — pulled live from payment_settings, so it reflects
        # the current admin-configured rate, not a snapshot.
        referralCommissionPercent: { type: integer }
        # Net lifetime referral commission in balance_transactions (kopecks):
        # referral credits minus referral_reversal rows from refunds/chargebacks.
        # Not tied to a billing period.
        referralEarningsKopecks: { type: integer }
        # Number of accounts registered using this user's referral code.
        referredUserCount: { type: integer }
        # Paying means at least one confirmed subscription payment.
        # Wallet top-ups do not qualify for referral progress or commission.
        referredPayingUserCount: { type: integer }
        # Host to build the shareable referral link against: the
        # admin-configured primary public domain if it's currently healthy,
        # otherwise the request's own (technical) host as a safety-net
        # fallback. Frontend must use this instead of window.location.origin.
        referralLinkHost: { type: string }
        # true while the account is administratively blocked. The frontend
        # must show a "banned" screen rather than redirecting to sign-in.
        isBanned: { type: boolean }
        # true when the user's current active subscription was never backed by a
        # completed payment — i.e. it is a free trial granted at registration.
        # Always false when hasActiveSubscription is false.
        isTrialSubscription: { type: boolean }
        # Opt-in automatic monthly renewal from the wallet balance.
        # When true the autoRenewJob debits the plan price ~24 h before expiry.
        autoRenewFromBalance: { type: boolean }
    ExpiredSubscription:
      type: object
      required: [id, endsAt, planName, billingType, isTrial]
      properties:
        id: { type: integer }
        endsAt: { type: ["string", "null"], format: date-time }
        planName: { type: string }
        billingType:
          $ref: "#/components/schemas/PlanBillingType"
        isTrial: { type: boolean }
    PlanBillingType:
      type: string
      enum: [monthly, hourly]
...
    AdminPasswordResetResult:
      type: object
      required: [resetUrl]
      properties:
        resetUrl: { type: string }
...